Refactor model initialization and memory management
- Renamed `xavier_seed` to `weight_init_seed` for clarity in model initialization.
- Updated logging to reflect the new weight initialization seed.
- Removed unnecessary training state allocations and validations, simplifying the model allocation state.
- Enhanced GPU encoder initialization to accept the new weight initialization seed.
- Introduced `CublasHandleOwner` and `DeviceAllocation` classes for better management of cuBLAS handles and CUDA device allocations.
- Added `GenerationState` for autoregressive inference state management.
- Updated VRAM breakdown documentation to reflect changes in memory usage and allocation.
- Adjusted training loop and diagnostic training scripts to accommodate new initialization parameters.

+++ resources/models/GRIM-text/Shared/Telemetry/TelemetryUpdate.cu
@@ -127,7 +127,7 @@ void populateExecBlockHealthStreams(
     float exec_active_ratio = 0.0f;
 
     if (gm.execution_block_count > 0) {
-        exec_grad_norm = std::sqrt(gm.execution_block_sum_sq / static_cast<float>(gm.execution_block_count));
+        exec_grad_norm = static_cast<float>(std::sqrt(gm.execution_block_sum_sq / static_cast<double>(gm.execution_block_count)));
         if (enc_rms_pre > 1e-12f) {
             exec_grad_ratio = exec_grad_norm / enc_rms_pre;
         }
